# contracts/snapshots/r4-go-f0f482c0/openapi.yaml
openapi: 3.0.3
info:
  title: Octo Meeting Service
  version: r4-go-f0f482c0
  description: >
    Service-owned, authoritative OpenAPI snapshot for the standalone Octo Meeting
    service. Bound to architecture SHA-256
    f0f482c0add1cfe81edb8cdab6418c63097be995037b582c339bbaeec0a8e871 and Go runtime
    revision 28e3712d1aa6b5db166067404770f90314ffbf3c45eaffa70a17a3f7e16ed723.


    This snapshot is the single source of truth for the public HTTP contract
    (architecture §7 / backend appendix §5). The frontend typed-client, MSW
    fixtures, and E2E reconcile against it. Wire format is snake_case; timestamps
    are UTC ISO-8601. The stable error taxonomy lives alongside in errors.yaml and
    is surfaced here as the error_response envelope.


    Handlers may be stubbed while this contract is published early to unblock the
    frontend; the shapes below are frozen regardless of handler readiness.
servers:
  - url: /meeting/api/v1
    description: Public gateway base path (browser / desktop entry).
  - url: /v1
    description: Canonical service path behind the gateway.
...
paths:
...
  /meetings/{meeting_id}/participants/{uid}/role:
    put:
      tags: [controls]
      operationId: setParticipantRole
      summary: Set a participant role (host only)
      description: The cohost may not assign or revoke cohost; only the host may.
      parameters:
        - $ref: "#/components/parameters/meeting_id"
        - $ref: "#/components/parameters/uid"
        - $ref: "#/components/parameters/x_space_id"
        - $ref: "#/components/parameters/if_match"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [C, M] }
      responses:
        "200":
          description: Updated participant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/participant" }
        default:
          $ref: "#/components/responses/error"
...
  /meetings/{meeting_id}/participants/{uid}:
    delete:
      tags: [controls]
      operationId: removeParticipant
      summary: Remove a participant (host/cohost)
      description: >
        Writes a terminal removal record that blocks future admission before the
        password step.
      parameters:
        - $ref: "#/components/parameters/meeting_id"
        - $ref: "#/components/parameters/uid"
        - $ref: "#/components/parameters/x_space_id"
        - $ref: "#/components/parameters/if_match_optional"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Removed participant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/participant" }
        default:
          $ref: "#/components/responses/error"
  /meetings/{meeting_id}/lock:
    put:
      tags: [controls]
      operationId: setLock
      summary: Lock or unlock the meeting (host/cohost)
      description: A lock rejects not-yet-in-room admissions; active users are unaffected.
      parameters:
        - $ref: "#/components/parameters/meeting_id"
        - $ref: "#/components/parameters/x_space_id"
        - $ref: "#/components/parameters/if_match"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [locked]
              properties:
                locked: { type: boolean }
      responses:
        "200":
          description: Lock state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/meeting_summary" }
        default:
          $ref: "#/components/responses/error"
...
  /meetings/{meeting_id}/share:
    post:
      tags: [controls]
      operationId: startShare
      summary: Start screen share
      description: Rejected with MEETING_SHARE_CONFLICT when a holder is already active.
      parameters:
        - $ref: "#/components/parameters/meeting_id"
        - $ref: "#/components/parameters/x_space_id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/share_request" }
      responses:
        "200":
          description: Share holder
          content:
            application/json:
              schema: { $ref: "#/components/schemas/share_state" }
        default:
          $ref: "#/components/responses/error"
    delete:
      tags: [controls]
      operationId: stopShare
      summary: Stop screen share
      parameters:
        - $ref: "#/components/parameters/meeting_id"
        - $ref: "#/components/parameters/x_space_id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/share_request" }
      responses:
        "200":
          description: Share state (empty holder)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/share_state" }
        default:
          $ref: "#/components/responses/error"
...
components:
  parameters:
    meeting_id:
      name: meeting_id
      in: path
      required: true
      schema: { type: string }
    uid:
      name: uid
      in: path
      required: true
      schema: { type: string }
    x_space_id:
      name: X-Space-Id
      in: header
      required: true
      description: Active Space id. Identity (uid/org) is resolved by the auth seam, never from client headers.
      schema: { type: string }
...
    if_match:
      name: If-Match
      in: header
      required: true
      description: Current entity version for optimistic concurrency.
      schema: { type: string }
    if_match_optional:
      name: If-Match
      in: header
      required: false
      schema: { type: string }
  responses:
    error:
      description: Stable error envelope (see errors.yaml for the code taxonomy).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/error_response" }
  schemas:
...
    error_response:
      type: object
      description: >
        Shared error envelope for all endpoints (snake_case). `code` is a stable
        value from errors.yaml; `details` carries only that code's safe keys and
        never a secret.
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
        request_id: { type: string }
...
    meeting_status:
      type: string
      enum: [scheduled, live, ended, cancelled]
    meeting_type:
      type: string
      enum: [quick, scheduled]
    role:
      type: string
      description: Host, Cohost, or Member.
      enum: [H, C, M]
    end_reason:
      type: string
      enum: [host_end, no_show, empty_timeout, system]
...
    meeting_summary:
      type: object
      required: [meeting_id, type, status, password_enabled, version]
      properties:
        meeting_id: { type: string }
        type: { $ref: "#/components/schemas/meeting_type" }
        status: { $ref: "#/components/schemas/meeting_status" }
        topic: { type: string }
        scheduled_start_at: { type: string, format: date-time }
        duration_minutes: { type: integer }
        actual_start_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        cancelled_at: { type: string, format: date-time }
        end_reason: { $ref: "#/components/schemas/end_reason" }
        locked: { type: boolean }
        password_enabled: { type: boolean }
        join_link_available: { type: boolean }
        version: { type: integer }
...
    participant:
      type: object
      required: [uid, role, aggregate_state]
      properties:
        uid: { type: string }
        role: { $ref: "#/components/schemas/role" }
        aggregate_state:
          type: string
          enum: [invited, joined, left, removed, superseded]
        removed: { type: boolean }
        first_joined_at: { type: string, format: date-time }
        last_left_at: { type: string, format: date-time }
        version: { type: integer }
...
    share_request:
      type: object
      required: [participant_segment_id, share_type]
      properties:
        participant_segment_id: { type: string }
        share_type: { type: string, enum: [screen, window, tab] }
    share_state:
      type: object
      properties:
        holder_uid: { type: string }
        share_type: { type: string, enum: [screen, window, tab] }
        active: { type: boolean }
...